utilities: use a different value for the weak pointer
Using T is correct, however ECL (including the last release) have issues with
that particular weak value. The problem is fixed in ECL, however before there
is no new release Flexichain is not usable on this implementation. Changing T
to "foo" which is equally good from the semantic perspective fixes the issue.